backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
apify-client
requests
anthropic
